Extract isValidAbsoluteUrl to zero deps util
